Fix async exceptions handling
Problem: resources cleanup does not occur when the main function is killed by
asynchronous exception.

I see two possible solutions for this problem:

1. Treat all async exceptions like sync ones.
Might be not good, because later we print them like application exception.

2. Just perform the cleanups in `finally` section or whatever (currently
those sections are unprotected).

In this commit I go in the first way as a temporal solution, want to learn
the repository author's opinion first.